Implement first "solid" voxel mode based in DDA traversing.
Some basic coloring in a dense grid, doesn't scale very well, but the
traversal is trivial and we can modify the data as much as we want.

There's quite a few warnings now I bumped the GLSL version to 130 because
of glTextureFetch -- I need to remove the now deprecated stuff in the
shader.