Added light shadows. Cleaned up voxel-filling functions.
I've started adding some basic functions to fill the voxel textures, for
now they're just very basic, but an authoring tool could be built around
these.

In terms of rendering, added basic directional lighting with shadows by
simply extracting the raymarching code to a function and doing the shadow
test using raymarching as well. This is starting to be expensive!.