Added multipass rendering, in preparation for implementing DOF.
Implemented a simple accumulation buffer by blending each new frame with
the running average of the previous frames (till the counter is reset when
we move the camera). For now this is used to add some per-pixel jittering
to do brute force antialiasing, but the idea is to implement lens effects
with this mechanism.