Add some fake voxel colors
There's no RGB8 format for imageStore so I had to pad it with an extra
alpha channel to make it RGBA8. This could be used for the occupancy flag
but I'm a bit wary that this would result in worse performance because we
do many occupancy accesses (traversal) for each color access (hit) and thus
making the occupancy texture fatter may not be a good idea.