WIP encoding a morton-based SVO
- Refactored dense voxelization code on the CPU to take different kinds of
voxel representations (dense grid, the former mode, and also now a simple
array of morton codes).

- The array of morton codes will be turned into an octree by the Svo class,
filling the array with all the morton codes corresponding to the intermediate
nodes (which I believe can be done efficiently bottom-up if the input leaf
array is sorted). We will also fill the ChildMask and VoxelDepth fields of
the morton codes at this point.