Add/Remove voxel tool
Building on the previous voxel selection framework, once we have the index
and normal of the selected voxel, the add/remove tools just use imageStore
to set the occupancy and color textures.

There's an important fix on dda.h on how we calculate the normals, since
there were glitches with the previous code.