Implement importance-sampled image based lighting.
This largely follows PBRT's implementation, where we take the original image
and calculate two CDF functions over the pixel intensities. These functions
are then used to map a pair of uniform random variables to the corresponding
uv coordinates to sample from, according to the pixel luminance distribution.

The image data is loaded and processed using OpenImageIO.