Improve renderer/user dialogs interaction.
When the renderer is working, the UI became very unresponsive because we're
hogging the GPU. I've implemented a couple of signals (begin/end user
interaction) which are forwarded to the GLWidget to pause the render loop
while a dialog (e.g. a color picker, or a file dialog) is open. This makes
UI a lot more responsive.