WIP sampling from emissive voxels
Emissive voxels are considered light sources, and thus we need to enumerate
and pick them randomly in the direct lighting estimation part of the
path tracer integrator.

Right now I'm keeping an explicit array with all indices for voxels which
have an emissive material associated, however keeping this array up to date
can be problematic: not a big deal if we're modifying voxels one by one,
but modifying the properties of a given material means N voxels might be
added or removed from the list at once. Sampling per material and then per
option could make this faster, but I don't think this would lead to a truly
uniform sampling distribution (unless we make the probabilities of sampling
each material proportional to the summed power of all the voxels they're
associated to).

There's a hack in the current commit where the emission is multiplied by
a large number. I seem to be needing this to really see the effects of the
lighting from voxels, compared to the effects of the environment light. It
may be a bug somewhere, or maybe the ranges of radiance truly are that
different.