Removing GLSL warnings.
Not sure how I had this working before, but we were injecting #pragma
directives right before the #version number, which the GLSL compiler
doesn't like. I'm instead removing the version from the shaders themselves
and adding it to the preprocessor stream in the compiler. Not a neat
solution, just a workaround, since ideally this should remain on each
shader, but I do need the #includes working in the shaders.